backend/requirements/base.txt
# GyanGrit — Base Requirements
# Installed in all environments (dev + prod)

Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg[binary]>=3.1
python-dotenv>=1.0
openpyxl>=3.1
boto3>=1.34
requests>=2.31         # HTTP client — used by Fast2SMS OTP service
whitenoise>=6.6        # Static file serving — required in MIDDLEWARE in base.py
ably>=2.0              # Real-time channels — competition rooms, chat rooms
django-unfold>=0.86      # Modern Django admin theme
PyJWT>=2.8                      # LiveKit JWT token generation
